Перенёс класс Dynaloader в отдельный файл
