+++ GraphModel/WindowsFormsApplication/WindowsFormsApplication.csproj
@@ -48,6 +48,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConcreteKeyboard.cs" />
+    <Compile Include="Dynaloader.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
